updated the code to break it up a bit, and connected it to a Supabase db, and got the register set up as well. Will make sure login, logout, and stripe sign up work tomorrow
