+++ .gitignore
@@ -1,1 +1,4 @@
 .env*
+.python-version
+.venv
+__pycache__
