--- init_db.py
@@ -1,7 +1,0 @@
-from app import app, db
-from sqlalchemy import inspect
-
-with app.app_context():
-    inspector = inspect(db.engine)
-    if not inspector.has_table('user'):
-        db.create_all()
