+++ requirements.txt
@@ -1,8 +1,24 @@
-Flask==2.2.2
-flask-cors==3.0.10
-Werkzeug==2.2.3
 boto3==1.20.24
-flask_sqlalchemy
-flask_jwt_extended
-python-dotenv==0.20.0
+botocore==1.23.54
+click==8.2.1
+Flask==2.2.2
+Flask-Cors==3.0.10
+Flask-JWT-Extended==4.7.1
+Flask-SQLAlchemy==3.0.3
+greenlet==3.2.3
+itsdangerous==2.2.0
+Jinja2==3.1.6
+jmespath==0.10.0
+legacy-cgi==2.6.3
+MarkupSafe==3.0.2
+psycopg2==2.9.10
 pydub==0.25.1
+PyJWT==2.10.1
+python-dateutil==2.9.0.post0
+python-dotenv==0.20.0
+s3transfer==0.5.2
+six==1.17.0
+SQLAlchemy==2.0.41
+typing_extensions==4.14.0
+urllib3==1.26.20
+Werkzeug==2.2.3
